practica-2-c2-main/practica-2-c2-main/requirements.txt
gunicorn==20.1.0
python-decouple==3.4
whitenoise==5.2.0
psycopg2==2.8.6
postgres==3.0.0
psycopg2-pool==1.1
alembic==1.6.1
aniso8601==9.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
db==0.1.1
cors==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.3.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
six==1.15.0
SQLAlchemy==1.3.23
tldextract==3.1.0
urllib3==1.26.4
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0